FIX if --model_dir is not None and it does not exist, create it
Previously, if --model_dir is set but it is not a valid path, it is overwritten by a new path.
However, logreg needs to know where model_dir is for its spawned subprocesses.
Therefore, logreg sets model_dir in its config files and we want model_dir to be created if it does not exist.